requirements.txt
# Core Framework
streamlit>=1.30.0
python-dotenv>=1.0.0

# Vector Database & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2

# LLM Providers
groq>=0.4.0
ollama>=0.1.6

# Web Scraping
playwright>=1.40.0

# PDF Parsing (Multi-Modal)
unstructured[all-docs]>=0.12.0
Pillow>=10.0.0

# Financial Data
yfinance>=0.2.36

# Utilities
uuid6>=2024.1.12
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
